refactor(dedupe): extract useCodeMirrorSingleLine from DomainCodeField and RegexCodeField
The two single-line CodeMirror editors shared ~105 identical lines covering
the full editor lifecycle: mount-once setup, Compartment management, reactive
effects for value / aria-invalid / theme, and scrollDOM tabindex. That logic
now lives in src/hooks/useCodeMirrorSingleLine.ts.

DomainCodeField and RegexCodeField are now thin wrappers that supply only their
field-specific language extensions (domainHighlighter vs. regexLanguage +
syntaxHighlighting) and theme factory (createDomainEditorTheme vs.
createRegexEditorTheme).

Files created:
  src/hooks/useCodeMirrorSingleLine.ts

Files updated (call sites):
  src/components/UI/DomainCodeField/DomainCodeField.tsx
  src/components/UI/RegexCodeField/RegexCodeField.tsx

Originates from painpoint #1 identified by the code-deduplicator agent.
