+++ src/components/UI/RegexCodeField/RegexCodeField.tsx
@@ -1,14 +1,5 @@
-import { useEffect, useRef } from 'react';
-import { useTheme } from 'next-themes';
-import { Compartment, EditorState } from '@codemirror/state';
-import {
-  EditorView,
-  keymap,
-  drawSelection,
-  placeholder as placeholderExtension,
-} from '@codemirror/view';
-import { defaultKeymap, history, historyKeymap } from '@codemirror/commands';
 import { syntaxHighlighting } from '@codemirror/language';
+import { useCodeMirrorSingleLine } from '@/hooks/useCodeMirrorSingleLine';
 import { regexLanguage, regexHighlightStyle } from './regexHighlight';
 import { createRegexEditorTheme } from './cmRegexTheme';
 
@@ -30,13 +21,14 @@ export interface RegexCodeFieldProps {
   testId?: string;
 }
 
+const REGEX_LANGUAGE_EXTENSIONS = [regexLanguage, syntaxHighlighting(regexHighlightStyle)];
+
 /**
  * Single-line CodeMirror editor that highlights a regular expression live while
  * preserving the accessibility and theming contract of a Radix `TextField`.
  *
- * Built as a slimmed variant of `JsonCodeEditor`: it mounts the editor once and
- * funnels reactive prop changes through dedicated effects and `Compartment`s.
- * Newlines are rejected so the value stays a single-line pattern.
+ * Thin wrapper around `useCodeMirrorSingleLine`: the only regex-specific
+ * details are the language/highlight extensions and `createRegexEditorTheme`.
  */
 export function RegexCodeField({
   value,
@@ -49,111 +41,17 @@ export function RegexCodeField({
   name,
   testId = 'regex-code-field',
 }: RegexCodeFieldProps) {
-  const { resolvedTheme } = useTheme();
-  const isDark = resolvedTheme === 'dark';
-
-  const containerRef = useRef<HTMLDivElement>(null);
-  const viewRef = useRef<EditorView | null>(null);
-
-  // Props the mount-once setup reads without re-running.
-  const onChangeRef = useRef(onChange);
-  const valueRef = useRef(value);
-  const idRef = useRef(id);
-  const placeholderRef = useRef(placeholder);
-  const ariaLabelRef = useRef(ariaLabel);
-  const describedByRef = useRef(describedById);
-  const hasErrorRef = useRef(hasError);
-  const isDarkRef = useRef(isDark);
-
-  const themeCompartment = useRef(new Compartment());
-  const ariaCompartment = useRef(new Compartment());
-
-  onChangeRef.current = onChange;
-
-  // Create the editor once. Reactive updates flow through the effects below.
-  useEffect(() => {
-    if (!containerRef.current) return;
-
-    const state = EditorState.create({
-      doc: valueRef.current,
-      extensions: [
-        regexLanguage,
-        syntaxHighlighting(regexHighlightStyle),
-        history(),
-        drawSelection(),
-        // Keep the value single-line: drop any transaction adding a line break.
-        EditorState.transactionFilter.of((tr) => (tr.newDoc.lines > 1 ? [] : tr)),
-        placeholderExtension(placeholderRef.current ?? ''),
-        EditorView.contentAttributes.of({
-          id: idRef.current ?? '',
-          'aria-label': ariaLabelRef.current,
-          'aria-describedby': describedByRef.current ?? '',
-          spellcheck: 'false',
-          autocapitalize: 'off',
-          autocorrect: 'off',
-          'data-1p-ignore': 'true',
-        }),
-        ariaCompartment.current.of(
-          EditorView.contentAttributes.of({
-            'aria-invalid': hasErrorRef.current ? 'true' : 'false',
-          }),
-        ),
-        themeCompartment.current.of(createRegexEditorTheme(isDarkRef.current)),
-        keymap.of([...defaultKeymap, ...historyKeymap]),
-        EditorView.updateListener.of((update) => {
-          if (update.docChanged) {
-            onChangeRef.current(update.state.doc.toString());
-          }
-        }),
-      ],
-    });
-
-    const view = new EditorView({ state, parent: containerRef.current });
-    viewRef.current = view;
-    // Make the (horizontally) scrollable viewport reachable by keyboard so a
-    // long pattern can be scrolled without a pointer (axe scrollable-region).
-    view.scrollDOM.setAttribute('tabindex', '0');
-
-    return () => {
-      view.destroy();
-      viewRef.current = null;
-    };
-    // Mount-once: subsequent prop changes are handled by the effects below.
-     
-  }, []);
-
-  // Sync externally driven value changes (form reset, edit of another rule).
-  useEffect(() => {
-    valueRef.current = value;
-    const view = viewRef.current;
-    if (!view) return;
-    const current = view.state.doc.toString();
-    if (value !== current) {
-      view.dispatch({ changes: { from: 0, to: current.length, insert: value } });
-    }
-  }, [value]);
-
-  useEffect(() => {
-    hasErrorRef.current = hasError;
-    const view = viewRef.current;
-    if (view) {
-      view.dispatch({
-        effects: ariaCompartment.current.reconfigure(
-          EditorView.contentAttributes.of({ 'aria-invalid': hasError ? 'true' : 'false' }),
-        ),
-      });
-    }
-  }, [hasError]);
-
-  useEffect(() => {
-    isDarkRef.current = isDark;
-    const view = viewRef.current;
-    if (view) {
-      view.dispatch({
-        effects: themeCompartment.current.reconfigure(createRegexEditorTheme(isDark)),
-      });
-    }
-  }, [isDark]);
+  const { containerRef } = useCodeMirrorSingleLine({
+    value,
+    onChange,
+    id,
+    placeholder,
+    ariaLabel,
+    hasError,
+    describedById,
+    languageExtensions: REGEX_LANGUAGE_EXTENSIONS,
+    createTheme: createRegexEditorTheme,
+  });
 
   return <div ref={containerRef} data-testid={testId} data-name={name} />;
 }
